requirements.txt
aiogram==2.23
aiosqlite==0.17.0
fuzzywuzzy==0.18.0
magic-filter==1.0.11
python-dotenv==1.0.1